+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v2.0.2 - 2025-11-XX
+
+### Fixed
+
+- [Admin Service] Correctly return open locks after adding namespace to locks
+- [Admin Service] Publish events improved error handling
+
 ## v2.0.1 - 2025-11-24
 
 ### Fixed